[Improve]Ingress uses different versions of API according to differen… (#2697)
* [Improve]Ingress uses different versions of API according to different K8s versions.

* Fix the review problem

* Complete the dependency package
